Make "kernaux_strncpy" signature compatible with C stdlib "strncpy"

+++ include/kernaux/stdlib.h
@@ -12,11 +12,7 @@ __attribute__((nonnull));
 
 size_t kernaux_strlen(const char *s) __attribute__((nonnull));
 
-char *kernaux_strncpy(
-    char *dest,
-    const char *src,
-    unsigned long slen
-)
+char *kernaux_strncpy(char *dest, const char *src, size_t slen)
 __attribute__((nonnull));
 
 void kernaux_itoa(int d, char *buf, int base) __attribute__((nonnull));
